UserManagement related APIs and business-logic

+++ src/main/java/com/niiazov/usermanagement/repositories/ProfileRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ProfileRepository extends JpaRepository<Profile, Long> {
+public interface ProfileRepository extends JpaRepository<Profile, Integer> {
 
-    Optional<Profile> findByUser_Id(Long userId);
+    Optional<Profile> findByUser_Id(Integer userId);
 }
