+++ src/main/java/com/niiazov/usermanagement/repositories/UserRepository.java
@@ -2,8 +2,10 @@ package com.niiazov.usermanagement.repositories;
 
 import com.niiazov.usermanagement.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
+
 }
